client/requirements.txt
websockets
pyperclip
pystray
Pillow
